+++ src/main/java/com/barlinc/unusual_prehistory/entity/ai/goals/StethacanthusAttackGoal.java
@@ -1,6 +1,8 @@
 package com.barlinc.unusual_prehistory.entity.ai.goals;
 
 import com.barlinc.unusual_prehistory.entity.Stethacanthus;
+import com.barlinc.unusual_prehistory.entity.utils.UP2Poses;
+import com.barlinc.unusual_prehistory.registry.UP2SoundEvents;
 import com.barlinc.unusual_prehistory.registry.tags.UP2EntityTags;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.LivingEntity;
@@ -27,23 +29,23 @@ public class StethacanthusAttackGoal extends AttackGoal {
         if (target != null && target.isInWater()) {
             stethacanthus.lookAt(target, 30F, 30F);
             stethacanthus.getLookControl().setLookAt(target, 30F, 30F);
-
             double distance = stethacanthus.distanceToSqr(target.getX(), target.getY(), target.getZ());
-            int attackState = stethacanthus.getAttackState();
 
             stethacanthus.getNavigation().moveTo(target, 1.5D);
             if (distance <= 4) {
                 stethacanthus.setAttackState(1);
             }
-            if (attackState == 1) {
+            if (stethacanthus.getAttackState() == 1) {
                 timer++;
+                if (timer == 1) stethacanthus.setPose(UP2Poses.BITING.get());
+                if (timer == 5) stethacanthus.playSound(UP2SoundEvents.STETHACANTHUS_BITE.get(), 0.7F, stethacanthus.getVoicePitch());
                 if (timer == 6) {
-                    if (stethacanthus.distanceTo(target) < getAttackReachSqr(target)) {
+                    if (stethacanthus.distanceTo(target) < this.getAttackReachSqr(target)) {
                         stethacanthus.doHurtTarget(target);
                         stethacanthus.swing(InteractionHand.MAIN_HAND);
                     }
                 }
-                if (timer >= 9) {
+                if (timer > 20) {
                     timer = 0;
                     stethacanthus.setAttackState(0);
                 }